+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+## 2024-04-15
+
+* Documentation added.
+* The naming of functions now more intuitive.
+* More drawing examples in main.cpp.
+* some unit test in main.cpp.
+* RGBA colors are separated into gfxDrawColors.
+
+
 ## 2024-04-10
 
 Initial commit with a `main.cpp` file that compiles the library and produces bmp files